Fix numerical precision and GPU performance issues
Numerical fixes:
- Promote GPU integration accumulators (d_combined_sum, d_rawsum) from
  float to double to prevent precision loss for large stacks (N>500)
- Fix GPU kappa-sigma to use double throughout mean/variance computation,
  matching the CPU path (integration.c)
- Promote CPU integrate_mean accumulator from float to double
- Use double for Moffat kernel normalization sum in both CPU and GPU
  star detection to eliminate systematic bias from float accumulation
- Fix DLT homography scaling fallback to use Frobenius norm instead of
  unstable max-element selection when H[8] ≈ 0
- Promote Lanczos CPU warp accumulators from float to double for better
  precision on high dynamic range pixels
- Keep calibration flat normalization divisor in double precision

Performance fixes:
- Cache Moffat kernel parameters to skip redundant per-frame constant
  memory uploads when alpha/beta are unchanged
- Pre-allocate GPU reduction scratch buffer for star detection, removing
  per-frame cudaMalloc + cudaStreamSynchronize + cudaFree cycle
- Switch pipeline D2H staging buffers (lum/conv/mask) from malloc to
  cudaMallocHost (pinned memory) for faster transfers
- Add star_detect_gpu_cleanup() to free persistent GPU resources

All 170 existing tests pass with no regressions.

+++ include/integration_gpu.h
@@ -70,8 +70,8 @@ extern "C" {
  * Not thread-safe; use one context per pipeline instance.
  */
 typedef struct IntegrationGpuCtx {
-    float  *d_combined_sum;                      /* W*H floats — survivor-weighted sum  */
-    float  *d_rawsum;                            /* W*H floats — unclipped sum          */
+    double *d_combined_sum;                      /* W*H doubles — survivor-weighted sum  */
+    double *d_rawsum;                            /* W*H doubles — unclipped sum          */
     int    *d_combined_count;                    /* W*H ints   — survivor count         */
     int    *d_rawcount;                          /* W*H ints   — valid (non-NaN) count  */
     float  *d_frames[INTEGRATION_GPU_MAX_BATCH]; /* batch device frame buffers          */
